refactor: add role form for use in both cases
