+++ src/components/Groups/GroupRoleForm.tsx
@@ -1,48 +1,27 @@
-import {
-  Card,
-  CardContent as MuiCardContent,
-  CardProps,
-  FormGroup,
-  styled,
-} from "@mui/material";
-import { Form, Formik, FormikHelpers, FormikProps } from "formik";
+import { FormikHelpers } from "formik";
 import { useState } from "react";
-import { ColorResult } from "react-color";
-import { useTranslation } from "react-i18next";
 import { toastVar } from "../../apollo/cache";
 import {
   RoleFragment,
   useCreateGroupRoleMutation,
   useUpdateGroupRoleMutation,
 } from "../../apollo/gen";
-import { FieldNames } from "../../constants/common.constants";
 import { DEFAULT_ROLE_COLOR } from "../../constants/role.constants";
 import { getRandomString } from "../../utils/common.utils";
-import ColorPicker from "../Shared/ColorPicker";
-import Flex from "../Shared/Flex";
-import PrimaryActionButton from "../Shared/PrimaryActionButton";
-import { TextField } from "../Shared/TextField";
+import RoleForm from "../Roles/RoleForm";
 
-const CardContent = styled(MuiCardContent)(() => ({
-  "&:last-child": {
-    paddingBottom: 18,
-  },
-}));
-
-interface Props extends CardProps {
+interface Props {
   editRole?: RoleFragment;
   groupId: number;
 }
 
-const GroupRoleForm = ({ editRole, groupId, ...cardProps }: Props) => {
+const GroupRoleForm = ({ editRole, groupId }: Props) => {
   const [color, setColor] = useState(
     editRole ? editRole.color : DEFAULT_ROLE_COLOR
   );
   const [colorPickerKey, setColorPickerKey] = useState("");
   const [createRole] = useCreateGroupRoleMutation();
   const [updateRole] = useUpdateGroupRoleMutation();
-
-  const { t } = useTranslation();
 
   const initialValues = {
     name: editRole ? editRole.name : "",
@@ -91,65 +70,15 @@ const GroupRoleForm = ({ editRole, groupId, ...cardProps }: Props) => {
     }
   };
 
-  const handleChangeComplete = (color: ColorResult) => setColor(color.hex);
-
-  const unsavedColorChange = () => {
-    if (!editRole) {
-      return false;
-    }
-    return editRole.color !== color;
-  };
-
-  const isSubmitButtonDisabled = ({
-    dirty,
-    isSubmitting,
-  }: FormikProps<typeof initialValues>) => {
-    if (isSubmitting) {
-      return true;
-    }
-    if (unsavedColorChange()) {
-      return false;
-    }
-    return !dirty;
-  };
-
   return (
-    <Card {...cardProps}>
-      <CardContent>
-        <Formik initialValues={initialValues} onSubmit={handleSubmit}>
-          {(formik) => (
-            <Form>
-              <FormGroup>
-                <TextField
-                  autoComplete="off"
-                  label={t("groups.form.name")}
-                  name={FieldNames.Name}
-                />
-
-                <ColorPicker
-                  color={color}
-                  key={colorPickerKey}
-                  label={t("roles.form.colorPickerLabel")}
-                  onChange={handleChangeComplete}
-                  sx={{ marginBottom: 1.25 }}
-                />
-              </FormGroup>
-
-              <Flex justifyContent="end">
-                <PrimaryActionButton
-                  disabled={isSubmitButtonDisabled(formik)}
-                  isLoading={formik.isSubmitting}
-                  sx={{ marginTop: 1.5 }}
-                  type="submit"
-                >
-                  {editRole ? t("actions.save") : t("actions.create")}
-                </PrimaryActionButton>
-              </Flex>
-            </Form>
-          )}
-        </Formik>
-      </CardContent>
-    </Card>
+    <RoleForm
+      color={color}
+      setColor={setColor}
+      colorPickerKey={colorPickerKey}
+      handleSubmit={handleSubmit}
+      initialValues={initialValues}
+      editRole={editRole}
+    />
   );
 };
