+++ src/components/Roles/ServerRoleForm.tsx
@@ -1,15 +1,6 @@
-import {
-  Card,
-  CardContent as MuiCardContent,
-  CardProps,
-  FormGroup,
-  styled,
-} from "@mui/material";
-import { Form, Formik, FormikHelpers, FormikProps } from "formik";
+import { FormikHelpers } from "formik";
 import produce from "immer";
 import { useState } from "react";
-import { ColorResult } from "react-color";
-import { useTranslation } from "react-i18next";
 import { toastVar } from "../../apollo/cache";
 import {
   CreateServerRoleInput,
@@ -19,33 +10,21 @@ import {
   useCreateServerRoleMutation,
   useUpdateServerRoleMutation,
 } from "../../apollo/gen";
-import { FieldNames } from "../../constants/common.constants";
 import { DEFAULT_ROLE_COLOR } from "../../constants/role.constants";
 import { getRandomString } from "../../utils/common.utils";
-import ColorPicker from "../Shared/ColorPicker";
-import Flex from "../Shared/Flex";
-import PrimaryActionButton from "../Shared/PrimaryActionButton";
-import { TextField } from "../Shared/TextField";
+import RoleForm from "./RoleForm";
 
-const CardContent = styled(MuiCardContent)(() => ({
-  "&:last-child": {
-    paddingBottom: 18,
-  },
-}));
-
-interface Props extends CardProps {
+interface Props {
   editRole?: RoleFragment;
 }
 
-const ServerRoleForm = ({ editRole, ...cardProps }: Props) => {
+const ServerRoleForm = ({ editRole }: Props) => {
   const [color, setColor] = useState(
     editRole ? editRole.color : DEFAULT_ROLE_COLOR
   );
   const [colorPickerKey, setColorPickerKey] = useState("");
   const [createRole] = useCreateServerRoleMutation();
   const [updateRole] = useUpdateServerRoleMutation();
-
-  const { t } = useTranslation();
 
   const initialValues = {
     name: editRole ? editRole.name : "",
@@ -112,65 +91,15 @@ const ServerRoleForm = ({ editRole, ...cardProps }: Props) => {
     }
   };
 
-  const handleChangeComplete = (color: ColorResult) => setColor(color.hex);
-
-  const unsavedColorChange = () => {
-    if (!editRole) {
-      return false;
-    }
-    return editRole.color !== color;
-  };
-
-  const isSubmitButtonDisabled = ({
-    dirty,
-    isSubmitting,
-  }: FormikProps<Omit<CreateServerRoleInput, "color">>) => {
-    if (isSubmitting) {
-      return true;
-    }
-    if (unsavedColorChange()) {
-      return false;
-    }
-    return !dirty;
-  };
-
   return (
-    <Card {...cardProps}>
-      <CardContent>
-        <Formik initialValues={initialValues} onSubmit={handleSubmit}>
-          {(formik) => (
-            <Form>
-              <FormGroup>
-                <TextField
-                  autoComplete="off"
-                  label={t("groups.form.name")}
-                  name={FieldNames.Name}
-                />
-
-                <ColorPicker
-                  color={color}
-                  key={colorPickerKey}
-                  label={t("roles.form.colorPickerLabel")}
-                  onChange={handleChangeComplete}
-                  sx={{ marginBottom: 1.25 }}
-                />
-              </FormGroup>
-
-              <Flex justifyContent="end">
-                <PrimaryActionButton
-                  disabled={isSubmitButtonDisabled(formik)}
-                  isLoading={formik.isSubmitting}
-                  sx={{ marginTop: 1.5 }}
-                  type="submit"
-                >
-                  {editRole ? t("actions.save") : t("actions.create")}
-                </PrimaryActionButton>
-              </Flex>
-            </Form>
-          )}
-        </Formik>
-      </CardContent>
-    </Card>
+    <RoleForm
+      color={color}
+      setColor={setColor}
+      colorPickerKey={colorPickerKey}
+      handleSubmit={handleSubmit}
+      initialValues={initialValues}
+      editRole={editRole}
+    />
   );
 };
